studyindexer/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
celery>=5.3.6
redis>=5.0.1
chromadb>=0.4.22
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers==2.2.2
torch>=2.0.0
transformers==4.35.2
huggingface-hub==0.19.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
unstructured>=0.10.30
pdf2image>=1.16.3
pytesseract>=0.3.10
python-docx>=1.0.1
PyPDF2>=3.0.0
docx2txt>=0.8 